+++ src/OpenTelemetry.Contrib.Extensions.AWSXRay/AWSXRayIdGenerator.cs
@@ -179,7 +179,7 @@ namespace OpenTelemetry.Contrib.Extensions.AWSXRay
             {
                 SamplingDecision.RecordAndSample => ActivitySamplingResult.AllDataAndRecorded,
                 SamplingDecision.RecordOnly => ActivitySamplingResult.AllData,
-                _ => ActivitySamplingResult.PropagationData
+                _ => ActivitySamplingResult.PropagationData,
             };
 
             if (activitySamplingResult != ActivitySamplingResult.PropagationData)